tools/data_collection/requirements.txt
# Data collection dependencies
requests>=2.31.0
pyyaml>=6.0
tqdm>=4.66.0
pillow>=10.0.0
